requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.0
huggingface-hub==0.24.7
HyperPyYAML==1.2.2
idna==3.9
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
optuna==4.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.3
pooch==1.8.2
primePy==1.3
protobuf==5.28.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
-e git+https://github.com/huggingface/speechbox.git@e7339dc021c8aa3047f824fb5c24b5b2c8197a76#egg=speechbox
speechbrain==1.0.1
SQLAlchemy==2.0.34
sympy==1.13.2
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.4.1
torchmetrics==1.4.2
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
yarl==1.11.1
zipp==3.20.2
